compiler speed improvement
