+++ docs/progress-notes/changelog.md
@@ -221,3 +221,6 @@ The compiler is now much more aggressive in where and how it expands macros and 
 - Several places where macros could be incorrectly executed more than once (possibly causing unwanted side effects) have been fixed.
 - Fixed bug in size calculation of non-inline stack arrays. Previous behavior was a compiler assert.
 - Correctly handle `mod` for unsigned numbers. Previous behavior was to treat all inputs as 32-bit signed integers.
+
+## V1.0 Revised constant propagation, speed improvements
+Improved error messages around macros
