+++ goalc/make/Tools.cpp
@@ -30,8 +30,11 @@ bool CompilerTool::needs_run(const ToolInput& task) {
 bool CompilerTool::run(const ToolInput& task) {
   // todo check inputs
   try {
-    m_compiler->run_front_end_on_string(
-        fmt::format("(asm-file \"{}\" :no-time-prints :color :write)", task.input.at(0)));
+    CompilationOptions options;
+    options.filename = task.input.at(0);
+    options.color = true;
+    options.write = true;
+    m_compiler->asm_file(options);
   } catch (std::exception& e) {
     fmt::print("Compilation failed: {}\n", e.what());
     return false;
